Image Metrics

A dataset containing results of a survey asking for opinions about the quality of a set of distorted binary images. The data is in "long" format. There were 12 undistorted images ('base_img' and 'base_name') and 15 versions of each distorted image ('img_no', and 'distorted'). The 'mean' is the mean opinion score for each image: the average rating each respondent gave the distorted image compared to the base image. The standard deviation ('std') is also reported. Several different quantitative binary image quality metrics are also included, with their name in the 'metric' column and th value of the comparison in the 'score' column.

data(imagemetrics)

Format

An object of class tbl_df (inherits from tbl, data.frame) with 1120 rows and 8 columns.

Examples

head(imagemetrics)

#>   base_img img_no   base_name     distorted      MOS       SD     metric
#> 1        1      1 image-1.png image-1-1.png 62.36364 18.26621     catsim
#> 2        1      1 image-1.png image-1-1.png 62.36364 18.26621 catsimrand
#> 3        1      1 image-1.png image-1-1.png 62.36364 18.26621  catsimacc
#> 4        1      1 image-1.png image-1-1.png 62.36364 18.26621   Accuracy
#> 5        1      1 image-1.png image-1-1.png 62.36364 18.26621    AdjRand
#> 6        1      1 image-1.png image-1-1.png 62.36364 18.26621      Cohen
#>       score
#> 1 0.6910774
#> 2 0.5283328
#> 3 0.8541659
#> 4 0.8789062
#> 5 0.5741829
#> 6 0.7612944

aggregate(score ~ metric, imagemetrics, FUN = mean)

#>        metric     score
#> 1    Accuracy 0.8493851
#> 2     AdjRand 0.4785270
#> 3      catsim 0.5077398
#> 4   catsimacc 0.7504173
#> 5  catsimjacc 0.4849682
#> 6  catsimrand 0.4177054
#> 7       Cohen 0.6171894
#> 8      cwssim 0.6712845
#> 9     Jaccard 0.5635011
#> 10     msssim 0.5569971

with(imagemetrics[imagemetrics$metric == "accuracy", ],
    plot(mean, score, xlab = "Accuracy", ylab = "Mean Opinion Score"))

#> Error in seq.int(from, to, length.out = n): 'from' must be a finite number